Remove utils and merge them to the modeling file
